fix login same address from extension to google
