+++ package.json
@@ -22,7 +22,7 @@
   },
   "dependencies": {
     "@polkadot/vue-identicon": "2.9.14",
-    "@sora-substrate/util": "1.15.11",
+    "@sora-substrate/util": "1.15.12",
     "@soramitsu/soramitsu-js-ui": "^1.0.45",
     "@subwallet/wallet-connect": "^0.2.3",
     "@urql/core": "^3.0.5",
